+++ skforecast/metrics/__init__.py
@@ -2,6 +2,5 @@ from .metrics import (
     mean_absolute_scaled_error,
     root_mean_squared_scaled_error,
     add_y_train_argument,
-    calculate_autocorrelation,
     _get_metric,
 )