+++ skforecast/plot/__init__.py
@@ -3,5 +3,6 @@ from .plot import (
     plot_multivariate_time_series_corr,
     plot_prediction_distribution,
     plot_prediction_intervals,
+    calculate_autocorrelation,
     set_dark_theme
 )